Adding new repositories for Octavia
The octavia project is adding the following repositories:
  octavia-dashboard
  octavia-tempest-plugin
  python-octaviaclient

This patch sets up those repositories and initial jobs.

octavia-dashboard will be pre-populated with a copy of
neutron-lbaas-dashboard that has the history preserved, but the
branches and tags have been removed.
The other repositories are fresh development and have no code to populate.
They will be setup using cookiecutter.

Change-Id: Id6385b8940fbef6c703db26808cda9607aea360d