Fix tempest-dsvm-neutron-multinode-full job VNC console settings
Commit https://github.com/openstack/tempest/commit/1f87a5611a added a new
tempest test "compute.servers.test_novnc.NoVNCConsoleTestJSON.test_novnc".
This test fails intermittently in multinode jobs. The test revealed that
the multinode test job has a wrong configuration.

The compute node ("subnode") in the multinode tests has a wrong value for
the config options:
* [vnc].vncserver_proxyclient_address
* [vnc].vncserver_listen
The defaults of these config options are "127.0.0.1", which makes sense in
an all-in-one test setup, but it should be the IP address of the compute node
in an multinode setup.

As the multinode test job uses a controller node (*with* compute services)
+ compute node, this also explains why this fails intermittently. The test
passes if the instance gets scheduled onto the controller node (*with*
compute service) and fails if the instance gets scheduled onto the compute
node ("subnode").

As the tempest tests and the novnc-proxy server are co-located on the
controller node, having "127.0.0.1" for "[vnc].novncproxy_base_url" is
fine(-ish). In a real-world setup this should be the IP address of the
node, where the novnc-proxy service is running (typically the controller).

The flow of commands is this:
* nova-api ask nova-compute for a valid VNC connection info object
  (with an "access_url")
* nova-consoleauth authorizes the connection info with a token
* nova-api returns the access_url + token
* a vnc client uses the access_url + token to establish a connection
* that access_url points to the nova-novncproxy service
* the nova-novncproxy service uses the token to get the connection info
  object from nova-consoleauth
* that connection info object contains the IP address of the compute node
  where the instance lives
* nova-novncproxy establishes a connection to the compute node
* the instance listens on the IP address of the compute node for incoming
  traffic

The "access_url" is derived from the config option (confusingly named)
vnc.novncproxy_base_url and the IP address the nova-novncproxy
uses is derived from the config option vnc.vncserver_proxyclient_address

The important code pieces are:
* nova-api:
  https://github.com/openstack/nova/blob/324768752f0a3a7e9ca14a57770556dbca419cc1/nova/compute/api.py#L3419-L3430
* nova-compute manager:
  https://github.com/openstack/nova/blob/63805735c25a54ad1b9b97e05080c1a6153d8e22/nova/compute/manager.py#L4543-L4571
* The console class:
  https://github.com/openstack/nova/blob/50c8f93e51fba12a154975d3d82c9cd34bc3b752/nova/console/type.py#L16-L28
* Libvirt driver:
  https://github.com/openstack/nova/blob/5f16ce546f808b826fd044229f236d232141551b/nova/virt/libvirt/driver.py#L2800-L2817

Co-Authored-By: Michelle Mandel <mmandel@us.ibm.com>
Change-Id: Idfc33d838d1d83793b7d9e4bc75acd4b14e622ab
Closes-Bug: 1669468