Add Swift tox functional EC env tests to gate
A new toxenv enviroment has been added to the Swift repo 'func-ec'
which will run the swift functional tests on an EC policy:
https://review.openstack.org/#/c/442749/

This patch added this toxenv as a check/gate job so we can be sure to
gate on EC better. Initially voting will be off just to be sure it's
working correctly.

As this toxenv only existing currently in master, and therefore only
future stable branches, current stable branches are filtered out.

Change-Id: I037c622e1e96d8790e0e9788e6485fcb06ab68a1