Print out network configuration data in net-info
There has been a few scenarios where eth1 doesn't get an IP
address which results in subnodes being unaccessible and
builds failing.

This patch prints out the information consumed by Glean in
order to help troubleshoot the issue and figure out the
root cause of the issue if it occurs again.

Change-Id: Ifd50e83fe3d73064b9141ed62380ec57277241d7