Run separate jobs for py2 and py3 d-g unittests
It turns out that our run-tox.sh scripts can't handle multiple tox envs
in the envlist. Instead of running d-g unittests under py2 and py3 in
the same job using multiple env list entries just run two jobs.

A future improvement to run tox to handle multiple envs in the envlist
would be nice but that affects far more jobs so this is simpler for now.

Change-Id: I00a7a71133f6bd8f2ec45c4a904892979f01ad36