Add a new project sushy-tools
This patch is adding a new project called "sushy-tools". This new
project will contain a set of tools to help with the development and
tests of the Sushy library.

Currently, these tools lives in the Sushy repository itself but they
have different requirements than the library itself (such as libvirt for
the simulator) so it makes sense to split it in another repository which
can be installed separately.

Needed-By: I880d9058408792f0c5d9f6531ae8ea06e06229ea
Change-Id: I2ff0299072e86dcf394e438ed2b6e443156a54e8