networking-bagpipe: add an experimental fullstack job
This is a first step to enable the fullstack
tests of I967ba442ef433b7ada5a1a381f1571ec3cf23517 in the gate.

It also adapts the scrip & template for the existing
tempest job so that 'tempest' is passed to gate_hook.sh in the
case of that job.

Change-Id: I29c5d3798f3a3eed48aef6bd744ea71486135c8e