Add experimental murano functional test gate.
This gate will be used for running murano-ci functional tests
for openstack/murano repository only. It is meant as a replacement
for murano-ci, hosted on Mirantis servers, so that infra can
maintain the gate.

Initially it will be run in the experimental pipeline only. Once
the gates pass, they will be promoted to non-voting/voting.

Change-Id: Ibd043dbc656d95d470b709c4bae0c96d3625106f