Add a DSVM job for golang projects
k8s-cloud-provider is an external driver for kubernetes. So
so to test code in this repo, we need to stand up a working
devstack environment, run etcd and kubernetes with proper
configuration and run kubernetes with the --cloud-provider
option pointing to our driver. Then we can run a bunch of
e2e tests say for example to test cinder integration. Any
repo that wants to use this job will need to have a devstack
plugin defined and the shell scripts for the hooks need
to be present in their proper location.

Change-Id: I03451816e2db66fbffc57a5de570f1505080bc0d