Remove enabled_services=tempest from patrole gate jobs.
Currently, ENABLED_SERVICES=tempest is contained in the gate jobs
for patrole. This appears to be causing problem when setting
up the multinode devstack environment [0], because

  tempest verify-config

is executed, which fails because:

  AttributeError: 'Manager' object has no attribute 'image_client'

which probably happens because tempest isn't set up on the other nodes,
causing tempest verify-config to fail.

Other multinode setups do not contain ENABLED_SERVICES=tempest:
for example [1]. Regardless, tempest is added to ENABLED_SERVICES
automatically by a script, apparently after devstack is done [1].

[0] http://logs.openstack.org/68/452768/2/check/gate-tempest-dsvm-patrole-multinode-admin-ubuntu-xenial-nv/6401f12/logs/devstack-subnodes-early.txt.gz
[1] http://logs.openstack.org/41/452041/7/check/gate-tempest-dsvm-neutron-scenario-multinode-ubuntu-xenial-nv/5907cf4/logs/devstack-subnodes-early.txt.gz

Change-Id: I0d59d6b956ff50e18364775054e16be79c39d799