Manage unbound on zuul workers in DIB
Currently, some parts of unbound are configured on zuul workers via
puppet and other parts are configured via dib elements.  As part of
reducing the nodepool image builds' dependency on puppet, this patch
moves the rest of what was done in the unbound puppet module to the
nodepool-base DIB element.

Note that while the puppet module explicitly calls for the defaults file
to exist before the package is installed[1], this does not apply for an
image built in a chroot where the service won't be started, so we can
stick that file in with the rest of the configuration.

[1] http://git.openstack.org/cgit/openstack-infra/puppet-unbound/tree/manifests/init.pp#n34

Change-Id: I3905be12acd85581a608d87ba5159cc883343a37