Enabling (Adding) Octavia to SDK gate.
Currently the fucntional tests written for Octavia controllers in
SDK are not able to hit the API endpoints for respective services.
So the tests are not validated properly. This patch would solve that
by enabling Octavia in SDK gates.
Also adding a local_conf section would enable other projects to add
for similar reasons.

Change-Id: Id64c4d8d33194332b121c16187428b1574144cc1