Write mirror host location if mirror is in use
In various places we have wanted to have just the mirror host location
and not specific mirror config like for pip or apt or yum. For example
devstack-gate attempts to use this information to do simple networking
ping tests, dib tests construct yum repo locations on ubuntu and apt
repo locations on centos, and so on. Have a clear file that is set to
the root mirror location if a mirror is in use to make this simpler.

Change-Id: I306467f1870e5ee823b8b5c98d81917e607348ef