Remove Tempest mitaka voting jobs
With the mitaka branch going EOL[1], we no longer need to run
mitaka jobs.

mitaka jobs are voting on check pipeline on tempest. To avoid
tempest gate block we need to remove those before stable repos are gone.

This patch only remove the check voting jobs and not periodic as those are
going to be cleanup with complete cleanup for mitaka EOL.

[1] https://releases.openstack.org/

Change-Id: Icc9a968472313de4d5cedcea06e733fdfa696003