Add nova-network functional job for openstackclient
We decide to support nova-network related commands in
python-openstackclient for stable openstack deployments,
sadly nova team mark nova-network as deprecated feature and
remove the support in novaclient 8.0.0. We rework to talk with
nova API directly bypass novaclient, now we need a job to verify
the changes, enable both n-net and n-cell, because n-net only can
be used in nova cell v1, and this is a non-voting job, because
some features are missing in cell v1, like: security group.

Change-Id: If4ef17b16f2fcf4698c035402b894ea660eaf976