Ignore retired repos for nodepool images
Do not clone repos that are retired. Since they will all use now
"retired.config" as file name, we can easily find them.

For this, move project filtering from main to get_project_list so that
we have access to the full yaml file and not only to the list of repo
names.

Change-Id: Iea6781c19b4828fd9bae0122add2f83e2586f190