Fix Tripleo upgrade tests by increasing timeout
It's currently hard to impossible to get anything merged if these tests
are voting (for example on the stable/ocata branch); please see the
related bugs. The original reason is not yet known, so this should be
temporary to get the gate passing again until the primary reason is
solved.

Related-Bug: 1684272
Related-Bug: 1680259

Change-Id: I6a470caccf88d7c3fff6e52d2a9b227e548d607d