Move magnum devstack job for shade to experimental
The magnum gate job has been broken for a few weeks in an inscrutable
manner. Additionally, magnum's own devstack jobs are vastly different
than shade's, so this is effectively a shade-specific setup. Remove for
now, just to stop wasting the nodes. It'll be in experimental should
anyone want to work on debugging it.

Change-Id: I92e8d82040c9ebd49d2410afe58c5c1282a7432c