n8g-bagpipe: use pre/post gate hooks to fullstack job
Add pre/post gate hoos to the fullstack job.

The hooks are added in master branch in change
Ic2bd4e1ab47bbd4c5e086ae6f5529e4ae50e74d8 which this change depends on, and
don't need to be added in stable branches for which this experimental
job won't be enabled.

Depends-On: Ic2bd4e1ab47bbd4c5e086ae6f5529e4ae50e74d8
Change-Id: I21f0ca6c92252961db5d4c44e95cd2c737164ed5