neutron: Add functional job to the gate queue
The job proven itself to be stabilized. At grafana dashboard [1], it's
observable that curve copies more or less the curve of the API job. Note
that this graph shows check queue which contains legitimate failures.

The patch adds only Xenial to the gate queue as Trusty job was not
worked on and will be abandoned after Mitaka EOL.

[1] http://grafana.openstack.org/dashboard/db/neutron-failure-rate?panelId=7&fullscreen

Change-Id: Id885a7cd6ae54d05b37965d2a6fa542575170c9e