Cleanup PROJECTS and DEVSTACK_PROJECT_FROM_GIT
Change Id2caf9561b361c1d4950856423282535d067e015 for devstack adds all
repositories in DEVSTACK_PROJECT_FROM_GIT to PROJECTS, so no need to add
repos in jobs to PROJECTS if there are in DEVSTACK_PROJECT_FROM_GIT.

This change went through all of jenkins/jobs and removed PROJECTS where
there is direct 1:1 correspondence. For a few files, especially ironic
and freezer, this was not doable.

Change-Id: I4b51e33b9bc0f1cb7948b69abf104e0b57b9ff52