manila: mark python-manilaclient to clone from git
Manila tests require the manilaclient from master, there for the
manilaclient is cloned in the manila devstack plugin.
Some projects (such as Fuxi, Karbor) depend on manila, and would
like to install the packaged manilaclient.
Align with other python clients - mark it to be cloned from git, and
in the devstack plugin clone it only if it has been marked.

Needed-By: Ia8e726b9358f6901c172a31f638a59eba523e0fa
Change-Id: I94d2f143159b60c4f8693e35fe844bc1e411e580