Disable EPEL after running Bindep for OSA jobs
EPEL is enabled on a per-transaction basis only when required as
part of the playbooks (ex: [1][2]).

By enabling EPEL for the purpose of installing packages out of
bindep and not disabling it afterwards means EPEL is effectively
"bleeding" into every yum transactions, not just the ones where it
is explicitely enabled.

This patch prevents such bleeding by disabling EPEL right after
bindep has finished installed packages.

[1]: https://github.com/openstack/openstack-ansible-lxc_hosts/blob/09406de86baed15c9e492d70538a96cb7035a9cc/vars/redhat-7.yml#L67
[2]: https://github.com/openstack/openstack-ansible-pip_install/blob/1db27456b9b6e57df3dc9d731e161ea71a3b6d9e/tasks/pre_install_yum.yml#L53

Change-Id: Iab82082d24f4edddeedd28964871dafc0379cd44