Create openstack-zuul-jobs / openstack-zuul-roles projects
As we continue with zuulv3, we are ready to start our first steps
towards writing jobs in ansible. As a result, we need a new
untrusted-project for zuul, which will allow dynamic reconfiguration
of jobs. We do not want to use project-config, as it is consider a
config-project and does not support dynamic reconfiguration.

Change-Id: I8732eba9303dc0748723f919d25d8c5dda3a05ed
Needed-By: Ie277f43a4032da24cccf008e8e3e66cc4c678a08
Signed-off-by: Paul Belanger <pabelanger@redhat.com>