Get rid of job name filter for ARA OSA job
In the original implementation, OSA uses a generic macro [1] that
is used by other projects and therefore needs to filter for their
own project as to not install EPEL on unrelated roles.

We copied the same shell builders to ARA's implementation but "name"
resolves to "ara", not "*openstack-ansible*" so that doesn't work.
It's irrelevant to filter on the project name regardless since this
job is only used inside the ARA gate so let's just get rid of it.

[1]: https://github.com/openstack-infra/project-config/blob/240c1e9696c73af35c452efbde3e502c92eb2473/jenkins/jobs/ansible-role-jobs.yaml#L79

Change-Id: I1c833ff245ea6033857615c915852618de4965be