networking-midonet: Introduce neutron-legacy jobs
Basically these jobs examine the case of neutron-legacy,
after switching the non-legacy jobs to use lib/neutron.
See networking-midonet gate_hook for the details. [1]

I didn't add Depends-On here because having the
experimental job first would help testing the
networking-midonet side of the change.

Once proven to work and the networking-midonet
change is merged, I plan to move this to non-voting,
along with the non-legacy counterpart of these jobs.

[1] I6883953d37e4c63010a412314abc6328eb996bd9

Related-Bug: #1666990
Change-Id: I487ad48ee9785e8a0579c945d11497c11531501c