Use master libs for shade functional-libs
The purpose of this job is to test that latest libs work with
newton devstack. We don't want to install devstack from branch
tips. Rather, we want to install branch tips (of master) into our
functional venv. This means we need zuul to prepare the repo states
for the libraries based on master, even though devstack is set with
a branch-override of stable/newton. It should be ok though, since
we did not set DEVSTACK_PROJECT_FROM_GIT for the projects as well.

Change-Id: I7de309504b5f52c1f7da64d3924efbb6db9a02e7