Fix senlin gate job for service enabling
Senlin API and functional tests do not require any other services other
than the fundational ones (rabbitmq, mysql, keystone, etc) to be
enabled. Senlin integration tests explicitly requires some services to
be installed and configured. This patch fixes both cases to make sure
that the services enabled are explicitly spelt out. This is crutial in
the situation that some services may happen to break our gate jobs and
we have to spend a lot time debugging the gate.

Change-Id: Ia1cd2c918420a04757cdd92a309b82bc830951f2