TripleO: container upgrade non-voting, not experimental
Promote container upgrade job from experimental to non-voting. The job
is now expected to be stable, and we'd like to prevent it from breaking
by running it on all tripleo-ci, t-h-t, tripleo-common and puppet-tripleo changes,
as non-voting currently.
Also do not run this job before Pike, since we don't support this
workflow.

Change-Id: Ia15827d0b68884261070603bfb2fd0226799b37e