Fix puppet-module-build
Change I082ea72743634b8ce0106786ab4ec35add874c8d updated the
job-template to use a list of nodes - and that broke the
gate-{name}-puppet-module-build template, since it used node as
variable. Looking at the job defintition of
gate-puppet-nova-puppet-module-build, it has:
<assignedNode>['centos-7', 'ubuntu-xenial']</assignedNode>

Explictely set the node parameter to the previous value of
centos-7.

Change-Id: I4c0e57c67273d0be0edf3bbd9d663a03144db5ae