Collapse shade legacy and neutron jobs
legacy and legacy-libs are only different in the git repo checkout.
neutron and neutron-libs are only different in the git repo checkout.

Collapse them from 4 definitions to 2, with parameters.

Also, collapse the template expansion into single calls. This will
expand a few combos we don't use, but that's not actually important
since we're not creating jobs in jenkins.

Finally, remove "neutron" from the normal job names. Neutron is the
default, there is no need to name it.

Change-Id: Ia1c5a25c8b3d1205e34456a46b862e3ea55cda43