Collapse shade legacy and non-legacy jobs
Similar to the last collapse, the differences in these sets of jobs is
small, and this gives us a very nice pile of combinations of things
available.

Also, remove "dsvm" from the name. These titles are getting a bit log.

There is a tiny hack, which is sed-ing /tmp/dg-local.conf after the
local_conf macro has written it - but that's just because there is one
local.conf setting we want different between legacy and non.

BTW - If you're wondering why I'm doing so much refactoring of shade's
jobs - it's because I'm starting to think about what they need to look
like in a zuul v3 world, so consolidating is the first step for my
brain.

Change-Id: Ie430cbb948de2ae9b5ee75d006d250588a44abbd