Reduce citycloud-la1 to max servers 0
This citycloud region does not have the 8C-8GB-100GB flavor so our boots
are failing. There are flavors with less disk and more disk available if
we keep 8 core and 8GB fixed or we can have 100GB of disk if we go up to
10GB of memory. Before we do that I want to ask them about possibly
adding in the 8C-8GB-100GB for consistency.

Change-Id: I32e405a950994b4e78940db579ab2760725e4391