Remove citycloud from multinode labels
There is a weird networking behavior in citycloud where the nova api
thinks a server has multiple private ip addresses. Nodepool picks one
and if it picks the one that is not configured on the host's network
interface private networking in multinode jobs does not work.

While we work with the cloud itself to help diagnose and correct this
issue stop running multinode jobs on citycloud to avoid failures related
to this problem.

Change-Id: I69ad48dcf272c6c77a41ac10d7e4ec187e96f39f