Revert "Remove citycloud from multinode labels"
This reverts commit 4a4d647292d84c252d05dc0782574dbc8dc368ef.

We have put a workaround in shade to pick the private ip associated with
the floating IP (which had to work if we are ssh'd to the node). This
means that even when there are multiple private IPs one of which is
valid we should always pick the valid one for multinode testing.
Therefore revert this change.

Change-Id: I5700a543fd30bd4ccbeadc14e7cdaa60526a1494