Fix ZUUL_USER_SSH_PUBLIC_KEY to support ssh key content
Currently ZUUL_USER_SSH_PUBLIC_KEY must be a key file, but
there is no change to create the key file, so it not convenient
for building a image with diskimage-builder, this patch makes
it possible to set ssh public key through ZUUL_USER_SSH_PUBLIC_KEY
in envirement(e.g. diskimage configuration in nodepool.yaml  ).

Change-Id: I83ef3d2e5ff5c3951ecf7036c46d0d8a68871e0c
Signed-off-by: shangxdy <shang.xiaodong@zte.com.cn>