bindep-fallbacks for openSUSE
Currently the bindep-fallback file assumes that platform:rpm
is only exectuted on Red Hat - family operating systems. On openSUSE
some of the package names are different (for legacy or consistency
reasons). Use platform:suse for matching SUSE package naming scheme.

mongodb and liberasurecode is currently missing from opensuse 42.2 (new in 42.3),
so let's see if we can get away without it. currently those are
blacklisted via !platform:opensuseproject and can hopefully be changed
to !platform:opensuseproject-42.2 once 42.3 is released and bindep
parsing bug related to the "." in the platform is fixed.

Change-Id: Iea28fb5a6805c4d015ce4d258fbf23153a45f49d