Explicitly clone requirements for bifrost
Bifrost expects to be able to have a complete folder on
disk for accessing the upper-constraint data, as opposed
to the single file on disk. Retooling to capture/use the
the single file will create more alternate paths that must
be taken to perform the install and it seems less ideal
in a deployed environment where the git repository in whole
provides some confidence of and history of the file.

Since we don't explicitly clone our own copy in OpenStack CI
we will have zuul cloner grab a copy for us.

Change-Id: I64d1744c776c20d906a29e072602c1909853da25