Make broken chef jobs experimental for now
The centos-7 and multinode jobs are consistenly broken, but take a
pretty long time to fail. Put them into the experimental queue so that
they can still continue to be improved, but don't hold up regular
patches too much.

Leave the ubuntu-xenial job in place, as there are patches in flight
that should make this one mostly working once we get the rubygems mirror
into action.

Change-Id: I906c489bbf5b49e6a65176e8c19e71e7ac08eaed