TripleO multinode container job
Add TripleO job for container deployment on the multinode
environment. See the depends-on change for the neede tripleo-quickstart
featureset.

This change also makes sure that all TripleO jobs that include
"container" in their name run only on Pike/master and later, since
that's when we started aiming at keeping containerized deployments
working all the time. The rule is consolidated so that we don't have to
introduce a new rule for each new container job.

Change-Id: I7d66d5614b2530c40f753802cdd9cf8920b13e0b
Depends-On: Iaa6166518de2943a540724af636f0e6be232e2ba