Move RBAC_TEST_ROLE env variable to local.conf
This is to move RBAC_TEST_ROLE={rbac-role} into the
localrc section of local.conf so the env variable
is properly set when devstack is running the patrole
plugin.

This commit also re-adds DEVSTACK_GATE_TEMPEST_ALL_PLUGINS
to the gate jobs because Patrole relies on site wide
packages for checking policy in code for (currently)
Nova and Keystone.

Change-Id: I3ff922f33590ae3b928eaa6b3c9c7f3d27038006