Add new repo for keystone tempest plugin
In addition to fulfilling Queens Goal "Split Tempest Plugins into Separate
Repos/Projects", this will allow us to re-enable the pbr's autodoc
building for the tempest plugin, which is currently disabled because the
oslo config option registration conflicts with the oslo configs from the
main project.

This imports from a temporary github repo that preserves the history of
the plugin and additionally applies the openstack cookiecutter to it.

Needed-By: Idd4eef295abf3f495559f7b70e98f2d7ce1050d2
Change-Id: I0fc551d38aa7243a02eed39b663a21e88f2ad713