Remove pypy jobs for python-cinderclient repos
Support for pypy has waned, and core components like the oslo libraries
are no longer tested against it. Jobs have been failing since we switched
to the cryptography library. Rather than look for a way around that, I
believe it is just time to remove pypy jobs.

This removes pypy for python-cinderclient and python-brick-cinderclient-ext
queues.

Change-Id: I4ae2ef1620ddd1b0f2b3a44ab13386980eba2364