diskimage-builder/glean: Upgrade nodepool-src-* jobs to voting
Recently a breakage in the nodepool-src jobs sneaked into
diskimage-builder git unintentionally. As those jobs are using
the -minimal elements and hence are not depending on 3rd party download
sites there shouldn't be instability caused by making those voting.

Also upgrade in glean so that there isn't a way for glean to break
diskimage-builder gate.

Change-Id: Iba7f787b4b637644a9f8829a4104b90c85e841f6