neutron: don't override projects to clone for functional job
We plan to use devstack-gate for the job (to inherit its setting
required for successful gate pass, like IPV4_ADDRS_SAFE_TO_USE). This
override makes it so that devstack-gate is not cloned, and hence an
attempt to trigger gate.sh script fails.

An alternative would be to add devstack-gate to the list, but Clark
Boylan told me we slowly move to removing all those overrides and just
trimming the default list of cloned projects, so it seems better to
align with infra ways, especially since we were bitten by the override
already.

Partial-Bug: #1693689
Needed-By: Iee7cc4c129e6e9eab91107fda92f66c5379d6f95
Change-Id: Ic874a08ce41dbfb2c7e43f61c92f1c7f13de16ea