Add scenario coverage to quickstart gates
We merged a bump to Ansible 2.3 which broke all of the scenario
jobs, because we did not have coverage for them in our gates. The
scenario jobs listed have file regexes, and so dont run on all
patches. We add the puppet jobs so that we ensure coverage in our
gate for all patches.

Change-Id: Ife6c335d58ddd367f2db87964476f85ddb46b617