Add seperate pep8 job for Browbeat
https://github.com/openstack/browbeat/blob/master/tox.ini

For some time the 'linters' job was running both the Ansible linting
script in the Browbeat tox.ini and the pep8 tests. Looking back I'm not
entierly sure how that ever worked. Anyways it was fixed recently so
I'm trying to add a new job for pep8. If this isn't the right way to
go about that please let me know.

Change-Id: I683b10beaedb49a0d8d2188817ce69f8df8ed921