New project - OpenStack-Spaceport
OpenStack-Spaceport is a project that aims to help deployment tools
especially, but not limited to, those focused upon containerised
environments. Building upon lessons learned in the development of
both Kolla-Kubernetes and OpenStack-Helm, OpenStack-Spaceport should
provide a common point of collaboration for across projects, providing
a suite of environment and dependency checks, that can be run prior to
launching an OpenStack or related service.

Change-Id: I996279edc3b73d0482b4600c3deaa52e63753a2d