Remove {pipeline}-cinder-dsvm-apache-{node}{suffix}
With the new switch to using uWSGI, we no longer need a separate
job to run Cinder with Apache. This removes the job completely,
including for stable branches pre-uWSGI, since it has never failed
and with stable policy it is highly unlikely we would change
anything there that could break it now. Easier just cleaning it
all up at this point.

Change-Id: I7b49c95e42ad0b7d120950ea59d4365be2305d8c
Depends-on: I8cef6c98f9242cc38d66de0ac499490e2a237887