Stop running python2 jobs for Zuul v3
We're adding things that require python3. Doing that in a manner that
still works for python2 is prohibitively difficult, given that they
actually use python2 language/syntax features.

We're running Zuul v3 with python3. There are no legacy installations of
v3. And the Software Factory folks say that running it with python3 on
centos7 is not an issue.

Go ahead and stop supporting it.

Change-Id: I7686df5493dd85d980390c9a0524419e19787d79