Make the ceph-plugin-src job non-voting for glance-store and os-brick
The gate-tempest-dsvm-full-devstack-plugin-ceph-ubuntu-xenial
job which runs against nova/cinder/glance/devstack/tempest changes
in the check queue has been non-voting for months now due to
instability and lack of resources to stabilize it.

When we made that job non-voting we failed to mirror that in the
ceph-plugin-src jobs that gate against os-brick and glance_storage
changes.

This makes the src job that runs against glance-store and os-brick
also non-voting.

Change-Id: I78b8b89e996f4213c63ffba689f5bdfc4a8772ed
Related-Bug: #1697953