Set PATH when running configure mirror script
The configure mirror script is trying to run utilities like ip and
restorecon which at least on some distros are hidden away in /usr/sbin
and not where nodepool would find them otherwise. Explicitly add /sbin,
/usr/sbin, /bin, /usr/bin, and /usr/local/bin to the end of PATH to list
all of the various places we might find things. With this in place we
should be able to find ip and restorecon whereever they are.

Change-Id: I800cd69e47b4b099b453e47f1b579ab015ed628f