Move dib image download jobs to -nv job
This moves the jobs that download images over the internet into a
separate -nv job.  We will work on getting the base images mirrored in
a similar fashion to how devstack does it, but it will take some time.

Now that nodepool has moved to Xenial builders, I think we can drop
the trusty-based testing.  I will follow-up on that shortly to reduce
the jobs here in the medium term.

Change-Id: Ib4dea95666143910b02109f17a9a0d217e9e10ba
Depends-On: Ie33ee49656872467ef68d753210032156bb6b2cb