Make infra's beaker jobs voting
Now that these jobs are in a good state, we need to make them voting in
order to prevent bitrot. This patch makes all of infira's puppet beaker
jobs voting except for puppet-infracloud, for which we instead re-add a
trusty job. Infracloud is still on Mitaka which isn't supported on
trusty, so we make sure it continues to work on trusty while we work on
migrating it to xenial and Newton/Ocata.

Change-Id: I015904ca1a65cf3c845e600c2d6061cd32f6da65