Co-gate spec helper gem with puppet-openstackci
In order to properly test changes in the spec helper gem that may affect
the puppet modules that depend on it, make sure the gem is checked out
from zuul as part of the puppet-openstackci beaker job. This job is run
for all infra modules so it is a reasonable job to co-gate with the
helper gem.

Change-Id: I66279da5f3cadb8a693c18379a2bdcf08edb2f15
Needed-by: I1b9ca7f38920a2c901ae16f6302af9b3a1ab36b6