promote all trove experimental scenario jobs to check as non-voting
Earlier changes made these jobs experimental as we focused our efforts
on the smaller subset of databases. Now as we begin to rebooth the
project and get all datastores back to proper operation, I'm promoting
these jobs to check (with non-voting) as a way of quickly identifying
the failure cases.

Change-Id: I4cb3aecd8abedac2055b4dc2d3a71d9728507769