diskimage-builder: remove suse from non-Xenial testing
This is a follow-on from too much copy-paste in
Ib4dea95666143910b02109f17a9a0d217e9e10ba.  As described in the
comments, opensuse-minimal can only be built on Xenial as it requires
zypper on the host, which trusty and centos don't have.

I've removed the python2/xenial templates as they're not required.
Rather than spin up another node for one test, let the nodepool jobs
test the python2/xenial/opensuse path.

Change-Id: Id76f128beef1253d970ef017e42a7b856799752b