jenkins: jobs: ansible-role-jobs: Always enable the EPEL repo
The ansible-role-jobs are supposed to be used in the OSA repositories so
there is not particular reason to check for the project name before we
enable the EPEL repository. If for whatever reason the EPEL repository
is not there, then we simply return 'true' so we don't break existing
jobs. This fixes a problem with the 'ansible-hardening' repository which
doesn't match the 'openstack-ansible' pattern and as such the EPEL
repository was never enabled for its jobs. This will soon affect the
ansible-hardening jobs with the following error:

2017-07-04 20:45:46.069562 | + sudo PATH=/usr/sbin:/sbin:/usr/local/bin:/usr/bin yum install -y python-ndg_httpsclient
2017-07-04 20:45:46.184220 | Loaded plugins: fastestmirror
2017-07-04 20:45:46.227381 | Loading mirror speeds from cached hostfile
2017-07-04 20:45:46.347841 | No package python-ndg_httpsclient available.
2017-07-04 20:45:46.435958 | Error: Nothing to do
2017-07-04 20:45:46.455212 | + set -e
2017-07-04 20:45:46.456459 | + try=3
2017-07-04 20:45:46.456522 | + /usr/bindep-env/bin/bindep -b -f bindep.txt test
2017-07-04 20:45:47.411528 | python-ndg_httpsclient
2017-07-04 20:45:47.423625 | + '[' 3 -gt 2 ']'
2017-07-04 20:45:47.423690 | + set +x
2017-07-04 20:45:48.408816 |
2017-07-04 20:45:48.408901 | ERROR: These requested packages were not installed:
2017-07-04 20:45:48.408911 |
2017-07-04 20:45:48.408925 | python-ndg_httpsclient
2017-07-04 20:45:48.408931 |
2017-07-04 20:45:48.408940 | + exit 1

Change-Id: If534b74d3e26f8f051e1c6186766d2e6434e21f9