Add publication job for service-types-authority data
Change the validate job to use the publish environment, since that
actually runs the full generation and local publication. Update
that to copy the generated files to the log server for examination,
both whatever is generated, and whatever publish thinks should be
published.

Add a job that publishes those results to the service-types directory on
static.openstack.org so that the service-types.openstack.org VHOST can
serve them.

Change-Id: I89e909f552afe967af9875799f599088957870cc