horizon: Run py27dj18 job for master
The default version of Django is now 1.10 in the master branch,
so we need to run py27dj18 rather than py27dj110.
py27dj110 is completely duplicated to python27 job.

On the other hand, we need to run py27dj110 job for stable branches
as the default django is 1.8.

The same change is made for django_openstack_auth.

Change-Id: I60efd6cf0a7dfbf7d8e2f688af855886cc59a8a8