Remove oslo.* from the oslo group
This was originally configured this way to make it possible to see all
of the bugs or blueprints for all of oslo in one place. However, it
seems unlikely that anyone is actually doing this anymore and it breaks
things like automagic linking of bugfixes to their corresponding
Launchpad bug. Simply remove the group.

Change-Id: I9746c05ffd6e7d62c84301dbbca5da3e9655785f