fix publishing directory depth for the unified publishing job
The run-docs.sh script already does some renaming and file moving for
us, so the unified publishing job needs to be more careful about the
renaming it does as it prepares files for publishing. We want everything
for a project to appear under docs.o.o/$project/, with master going to
"latest", other stable branches going to $series, and tagged versions
going to a directory named for the tag.

Change-Id: I4c0350b29ff910e5acdd1c0982d0096a20188495
Signed-off-by: Doug Hellmann <doug@doughellmann.com>