Do some openstack-ansible job housekeeping
The following experimental jobs have been removed
as their purpose has been served and are no longer
required.
- gate-openstack-ansible-openstack-ansible-compute-ubuntu-xenial-nv
- gate-openstack-ansible-openstack-ansible-ironic-ubuntu-xenial-nv
- gate-openstack-ansible-openstack-ansible-magnum-ubuntu-xenial-nv
- gate-openstack-ansible-openstack-ansible-sahara-ubuntu-xenial-nv
- gate-openstack-ansible-openstack-ansible-swift-ubuntu-xenial-nv

The following jobs have been promoted to voting
as they have proven reliable:
- gate-openstack-ansible-openstack-ansible-ceph-ubuntu-xenial
- gate-openstack-ansible-os_neutron-ansible-calico-ubuntu-xenial

The following job is being demoted as nothing has
been done to resolve the consistent failures. The
intent to do so is there, but we'd rather not waste
resources continually executing it for every commit.
- gate-openstack-ansible-os_neutron-ansible-dragonflow-ubuntu-xenial-nv

Change-Id: I78ed823909d601e464de0326dc333a6d75fa9ab3